backend/requirements.txt
# ─── Core API ───────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.7.0,<3.0
httpx>=0.27.0

# ─── LangChain 0.3 (Modern) ─────────────────────────────────────────────────
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0

# ─── Vector Store ───────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ─── Ingestion ──────────────────────────────────────────────────────────────
feedparser>=6.0.11
requests>=2.31.0
beautifulsoup4>=4.12.3

# ─── CrewAI (Modern) ────────────────────────────────────────────────────────
crewai>=0.102.0

# ─── Utilities ──────────────────────────────────────────────────────────────
numpy>=2.0.0
